event setting

+++ app/Http/Controllers/Setup/EventSettingController.php
@@ -20,7 +20,7 @@ class EventSettingController extends Controller
     }
 
     public function create(){  
-        $present_lists = Present::all();
+        $present_lists = Present::where('status',1)->get();
         $products = Product::all();
         return view('setup/event-setting-create', ['present_lists'=>$present_lists, 'products'=>$products]);
     }
@@ -34,7 +34,15 @@ class EventSettingController extends Controller
         $event_end = $request->input('end_time');
         $products = $request->input('product');
         $presents = $request->input('present_id');
-        $draw_probability = $request->input('draw_probability');
+        $draw_probability = array_filter($request->input('draw_probability'));
+        if(count($draw_probability) != count($presents)){
+            Alert::warning("The probability and presents count is did not match.")->persistent('Dismiss');
+                 return redirect()->route('event-setting-create');
+        }
+        if($event_start > $event_end){
+            Alert::warning('The start time cannot greater than end time.')->persistent('Dismiss');
+            return redirect()->route('event-setting-create');
+        }
 
         $products = implode(',', $products);
 
@@ -61,16 +69,17 @@ class EventSettingController extends Controller
                     ->whereDate('event_end_time', '=', date('Y-m-d', strtotime($event_end)))
                     ->get();
 
-
+                    $i=0;
                 foreach ($presents as $k => $value) {
                 
                     $eventDetails = new EventSettingDetail();
 
                     $eventDetails->event_id     = $event[0]->id;
                     $eventDetails->present_id   = $value;
-                    $eventDetails->present_prob = $draw_probability[$value -1];
+                    $eventDetails->present_prob = $draw_probability[$i];
                     $eventDetails->created_by   = Auth::user()->id;
                     $eventDetails->save();
+                    $i++;
                 }
             } else {
                 // if save is not successful
@@ -79,19 +88,21 @@ class EventSettingController extends Controller
             return redirect()->route('event-setting-index');
         } else {
             Alert::warning('Event Time is duplicate (တူညီသော အချိန်တွင် တခြား Event ရှိနေသောကြောင့် အချိန်ကို ပြန်ရွေးပါ')->persistent('Dismiss');
-            return redirect()->route('event-setting-index');
+            return redirect()->route('event-setting-create');
         }
     }
 
     public function overview($id)
     {
-        $present_lists = Present::all();
+        $present_lists = Present::where('status',1)->get();
         $products = Product::all();
         $res = DB::table('event_settings as es')
-        ->join('event_setting_details as esd','es.id','=','esd.event_id')
-        ->join('presents as pre','pre.id','=','esd.present_id')
-        ->join('products as pro','pro.id','=','es.product_id')
-        ->select('es.*','esd.*','pre.id as present_id','pre.present_name','pre.present_code','pro.id as product_id','pro.p_name as product_name','pre.image')->where('es.id',$id)->get();
+        ->leftJoin('event_setting_details as esd','es.id','=','esd.event_id')
+        ->leftJoin('presents as pre','pre.id','=','esd.present_id')
+        ->rightJoin('products as pro','pro.id','=','es.product_id')
+        ->select('es.*','esd.*','pre.id as present_id','pre.present_name','pre.present_code','pro.id as product_id','pro.p_name as product_name','pre.image')->where('es.id',$id)->where('pre.status','1')->get();
+      //  echo $res;die;
+
         $event_product = DB::table('event_settings')->where('id',$id)->get();
         $productID = $event_product[0]->product_id;
         $productID = explode(",",$productID);
@@ -100,7 +111,7 @@ class EventSettingController extends Controller
     }
     public function edit($id)
     {
-        $present_lists = Present::all();
+        $present_lists = Present::where('status',1)->get();
         $products = Product::all();
         $res = DB::table('event_settings as es')
         ->join('event_setting_details as esd','es.id','=','esd.event_id')
@@ -115,15 +126,29 @@ class EventSettingController extends Controller
     }
     public function update(Request $request)
     {
-        // dd($request->all());
-        DB::table("event_setting_details")->where('event_id',$request->input('event_id'))->delete();
-        DB::table("event_settings")->where('id',$request->input('event_id'))->delete();
+        // Validate the sum of percentage
+        if(array_sum($request->input('draw_probability'))!=100){
+            Alert::warning("The sum of percentage is must be 100%;")->persistent('Dismiss');
+            return redirect()->route('event-setting-edit',['id'=>$request->input('event_id')]);
+        }
 
         $event_start = $request->input('start_time');
         $event_end = $request->input('end_time');
         $products = $request->input('product');
         $presents = $request->input('present_id');
-        $draw_probability = $request->input('draw_probability');
+        $draw_probability = array_filter($request->input('draw_probability'));
+        // match percentage and present
+        if(count($draw_probability) != count($presents)){
+            Alert::warning("The probability and presents count is did not match.")->persistent('Dismiss');
+                 return redirect()->route('event-setting-edit',['id'=>$request->input('event_id')]);
+        }
+        // Validate date
+        if($event_start > $event_end){
+            Alert::warning('The start time cannot greater than end time.')->persistent('Dismiss');
+            return redirect()->route('event-setting-edit',['id'=>$request->input('event_id')]);
+        }
+        DB::table("event_setting_details")->where('event_id',$request->input('event_id'))->delete();
+        DB::table("event_settings")->where('id',$request->input('event_id'))->delete();
 
         $products = implode(',', $products);
 
@@ -143,7 +168,7 @@ class EventSettingController extends Controller
                 ->get();
 
                 // dd($event);
-
+                $i=0;
             foreach ($presents as $k => $value) {
                 $eventDetails = new EventSettingDetail();
 
@@ -152,22 +177,23 @@ class EventSettingController extends Controller
                 $eventDetails->present_prob = $draw_probability[$k];
                 $eventDetails->created_by   = Auth::user()->id;
                 $eventDetails->save();
+                $i++;
             }
         } else {
             // if save is not successful
         }
-        return redirect()->route('event-setting');
+        return redirect()->route('event-setting-index');
     }
     public function search(Request $request)
     {
         $searchTerm = '%' . $request->input('search') . '%';
 
-        $e_settings = DB::table('event_settings')->where('name','like',$searchTerm)->paginate(10);
+        $e_settings = DB::table('event_settings')->where('name','like',$searchTerm)->where('status',1)->paginate(10);
         return view('setup/event-setting', ['e_settings'=>$e_settings]);
     }
     public function delete($id)
     {
-         DB::table('event_setting_details')
+         DB::table('event_settings')
     ->updateOrInsert(['id' => $id], ['status' => 0]);
             return redirect()->route('event-setting-index');
